#169 The user can like or dislike a post
